Refactor desktop workflow: update relative paths for consistency, integrate event bus to sync chat list updates, and modify Sidebar/MenuChatList components to use new storage constants.
