+++ packages/desktop-workflow/src/helpers/HelperConstantsElectron.ts
